Refactor GUI & decouple from config
